python/requirements.txt
# Provider System Requirements

# Core Dependencies
fastapi>=0.104.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0

# Security
cryptography>=41.0.0

# HTTP Clients
aiohttp>=3.9.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
ruff>=0.1.0
mypy>=1.7.0
black>=23.0.0